Marking a class as beta, which in this rare set of circumstances is accurate.
This is a minor bug in ApiGenerator, which combined with a convoluted set of reproduction steps, resulted in CompareAgainstApiSpecs test blocking a change that should be legit.
Specifically, the removal of GetActionsOnPlayersInSegmentTaskStatusResult, which IS beta, and is the only server reference to ActionsOnPlayersInSegmentTaskSummary.
(Admin has a non beta reference to ActionsOnPlayersInSegmentTaskSummary, which makes it not beta here, which is the bug)
This checkin lets the removal of the beta APIs pass the tests, which suffices for the immediate unblock.